backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
psycopg2-binary
pydantic-settings
httpx
alembic
python-jose
PyJWT
email-validator
